Break up importxmlfirstpass
This file didn't have a very relevant name
