+++ src/importexport/musicxml/internal/musicxml/musicxml.cmake
@@ -15,11 +15,11 @@ set (MUSICXML_SRC
     ${CMAKE_CURRENT_LIST_DIR}/importmxmlpass2.cpp
     ${CMAKE_CURRENT_LIST_DIR}/importmxmlpass2.h
     ${CMAKE_CURRENT_LIST_DIR}/importxml.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importxmlfirstpass.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/importxmlfirstpass.h
     ${CMAKE_CURRENT_LIST_DIR}/musicxml.h
     ${CMAKE_CURRENT_LIST_DIR}/musicxmlfonthandler.cpp
     ${CMAKE_CURRENT_LIST_DIR}/musicxmlfonthandler.h
+    ${CMAKE_CURRENT_LIST_DIR}/musicxmlpart.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/musicxmlpart.h
     ${CMAKE_CURRENT_LIST_DIR}/musicxmlsupport.cpp
     ${CMAKE_CURRENT_LIST_DIR}/musicxmlsupport.h
     ${CMAKE_CURRENT_LIST_DIR}/musicxmltupletstate.cpp
