+++ src/importexport/musicxml/internal/musicxml/musicxmltypes.cpp
@@ -103,3 +103,119 @@ String VoiceDesc::toString() const
     }
     return res;
 }
+
+//---------------------------------------------------------
+//   interval
+//---------------------------------------------------------
+
+Interval MusicXmlIntervalList::interval(const Fraction f) const
+{
+    if (empty()) {
+        return {};
+    }
+    auto i = upper_bound(f);
+    if (i == begin()) {
+        return {};
+    }
+    --i;
+    return i->second;
+}
+
+//---------------------------------------------------------
+//   instrument
+//---------------------------------------------------------
+
+const String MusicXmlInstrList::instrument(const Fraction f) const
+{
+    if (empty()) {
+        return String();
+    }
+    auto i = upper_bound(f);
+    if (i == begin()) {
+        return String();
+    }
+    --i;
+    return i->second;
+}
+
+//---------------------------------------------------------
+//   setInstrument
+//---------------------------------------------------------
+
+void MusicXmlInstrList::setInstrument(const String instr, const Fraction f)
+{
+    // TODO determine how to handle multiple instrument changes at the same time
+    // current implementation keeps the first one
+    if (!insert({ f, instr }).second) {
+        LOGD() << "element already exists, instr: " << instr
+               << ", tick: " << f.toString() << "(" << f.ticks() << ")";
+    }
+    //(*this)[f] = instr;
+}
+
+//---------------------------------------------------------
+//   LyricNumberHandler
+//   collect lyric numbering information and determine order
+//
+//   MusicXML lyrics may contain name and number attributes,
+//   plus position information (typically default-y).
+//   Name and number are simply tokens with no specified usage.
+//   Default-y cannot easily be used to determine the lyrics
+//   line, as it tends to differ per system depending on the
+//   actual notes present.
+//
+//   Simply collecting all possible lyric number attributes
+//   within a MusicXML part and assigning lyrics position
+//   based on alphabetically sorting works well for all
+//   common MusicXML files.
+//---------------------------------------------------------
+
+//---------------------------------------------------------
+//   addNumber
+//---------------------------------------------------------
+
+void LyricNumberHandler::addNumber(const String& number)
+{
+    if (m_numberToNo.find(number) == m_numberToNo.end()) {
+        m_numberToNo[number] = -1;           // unassigned
+    }
+}
+
+//---------------------------------------------------------
+//   toString
+//---------------------------------------------------------
+
+String LyricNumberHandler::toString() const
+{
+    String res;
+    for (const auto& p : m_numberToNo) {
+        if (!res.isEmpty()) {
+            res += u" ";
+        }
+        res += String(u"%1:%2").arg(p.first).arg(p.second);
+    }
+    return res;
+}
+
+//---------------------------------------------------------
+//   getLyricNo
+//---------------------------------------------------------
+
+int LyricNumberHandler::getLyricNo(const String& number) const
+{
+    const auto it = m_numberToNo.find(number);
+    return it == m_numberToNo.end() ? 0 : it->second;
+}
+
+//---------------------------------------------------------
+//   determineLyricNos
+//---------------------------------------------------------
+
+void LyricNumberHandler::determineLyricNos()
+{
+    int i = 0;
+    for (auto& p : m_numberToNo) {
+        p.second = i;
+        ++i;
+    }
+}
